Use 1 minute rate for the metric timer
Making the metric function run with a 1 minute rate solves the
INSUFFICIENT_DATA problem, because we'll be sure to have a metric ready
when the alarm checks for it. It also decreases the lead time that the
ideal metric has over our retroactive calculation, speeding up the
scaling reaction time.